ore: add enhanced panic handler
This commit adds a method to the `ore::panic` module that installs a
panic hook that enhances the message provided out of the box by the Rust
runtime.

The most important enhancement is that the message is printed while
holding the stderr lock, which should prevent the message from being
badly interleaved with concurrent log output from the `tracing` crate.
Care is taken to fall back to outputting directly to stderr in case a
thread has wedged while holding the stderr lock.

A secondary enhancement is that backtraces are always collected and
printed even if `RUST_BACKTRACE` has not been set. The idea is that
panics are always unexpected and sometimes rare, and we don't want to
miss our chance to print the cause, even if it can be a bit more
verbose.

The abort-on-panic behavior of the existing `set_abort_on_panic` method
in the module is also folded into the new `install_enhanced_handler`
method.

Fix #24088.