compute: move `Instance` controllers into tasks
This commit moves the `Instance` controllers out of the coordinator task
into their own tasks. The `ComputeController` now doesn't hold on to
`Instance` objects directly, but to `InstanceState` handles that are
connected to `Instance`s through channels. `InstanceState` provides
methods to invoke the connected `Instance` through the channel,
optionally awaiting a result.

The conversion performed here is mostly mechanical and doesn't actually
improve things much: Most compute controller methods are now `async` and
block on responses from the `Instance`s they invoke, so the compute
controller isn't actually more decoupled than before. Subsequent commits
will concern themselves with making most of the compute controller
methods non-blocking.

Note that the code does not compile in this state, as the callers of the
compute controller have not been adapted to handle the `async` API. This
is intentional as most of the asyncness will be removed again.