catalog: Combine epoch and deploy generation
Previously, the catalog used both the epoch and the deploy generation
as fencing tokens. However, they were both written to the catalog as
separate entries and at different times. This complicated the fencing
logic and often resulted in fencing errors with less information than
we wanted.

This commit combines the epoch and deploy generation into a single
fencing token that is always written to the catalog atomically. It
utilizes the `FenceToken` update kind which was added in a previous
commit. Since the epoch and fence token must be available before the
catalog is opened (i.e. before migrations are run), this commit needs
to handle both epochs and fence tokens. In the next release we can
fully remove the `Epoch` update kind.

Additionally, fence errors can indicate with certainty whether the
fence is due to the deploy generation or epoch. Therefore, we can
remove the halt varint of `unwrap_or_terminate` which was used when
we weren't sure if the deploy generation had increased.