sql-parser: align SUBSCRIBE syntax with CREATE SINK/SELECT
The "output options" for `SUBSCRIBE` (i.e., the `ENVELOPE` and `WITHIN
TIMESTAMP ORDER BY` clauses) were jammed in at the end of the statement,
as in:

    SUBSCRIBE t WITH (SNAPSHOT) AS OF 1 UP TO 2 ENVELOPE UPSERT KEY (a)

This was inconsistent with our other SQL statements. In particular,
CREATE SOURCE and CREATE SINK always put the WITH options at the very
end of the statement:

    CREATE SINK ... WITH (SNAPSHOT)

And SELECT always puts the AS OF clause at the very end of the
statement:

    SELECT * FROM t WHERE ... AS OF 1

This commit restores the consistency that used to be present in
`SUBSCRIBE` by moving the output options to *before* the `AS OF`, `UP
TO`, and `WITH` options:

    SUBSCRIBE t ENVELOPE UPSERT KEY (a) WITH (SNAPSHOT) AS OF 1 UP TO 2

The old syntax remains supported for backwards compatibility, but is
softly deprecated. Only the new syntax will be mentioned in the docs
going forward.