storage: truncate replica status history
This commit adds support for truncation of the `ReplicaStatusHistory`
collection during storage controller startup.

In contrast to the other status histories, the replica status history is
keyed by two columns (`replica_id`, `process_id`), so we need to modify
the truncation code to work with multi-column keys. To this end, a
`StatusHistoryDesc` type is introduced that allows the caller to
configure the truncation behavior while being generic over the key type.