sql: Fix SHOW ROLE MEMBERSHIP filter (#29375)
The documentation for `SHOW ROLE MEMBERSHIP` says that the `FOR
<role_name>` filter does the following:

> Only shows role memberships granted directly or indirectly to
> role_name.

The implementation outputs these role memberships, but it also outputs
role memberships where other roles are members of role_name. This makes
the output confusing and not consistent with the documentation. This
commit updates the implementation so that it's consistent with the
documentation.