catalog: Optimize durable transaction operations
This commit provides two optimizations to durable catalog transactions.

The first is to return early from functions that update many objects,
but are provided an empty collection. For example, if `update_items` is
passed an empty map. These functions can be O(n^2) on the number of
objects in the catalog. So returning early for no-ops can save a lot of
time.

The second optimization is to make `uniqueness_violation` an `Option`.
`uniqueness_violation` is a function that is called on every pair of
objects during the `verify` method. Often `uniqueness_violation` is
hard coded as a function that always returns `false`. So we'd often
uselessly iterate over all pairs of objects (which is O(n^2)), doing
nothing. Now we only iterate over the objects when
`uniqueness_violation` is `Some`.

Works towards resolving #28722