catalog: Optimize uniqueness violation check
When updating an in-progress durable catalog transaction, we need to
verify that we're not violating the uniqueness constraint, if one
exists. Previously, we would modify the transaction and then check all
items against every other item so see if anything violates the
constraint. This is O(n^2) and can be fairly slow when there are a lot
of items. Additionally, the catalog changes are usually small and so
it's wasteful to always compare every item against every other item.

This commit adds a new method that only checks the newly inserted
items against all other items. This is O(n * k), where n is the number
of items and k is the number of new items. This commit also updates all
callers of the old method to switch to the new method.

Works towards resolving #28722