persist: add dangerous_force_compaction_and_break_pushdown method
As part of the effort to get envd startup times down, we've recent
written retractions for every storage usage entry in the catalog shard.
This has, ironically, made them all twice as big. Our primary plan
remains "hope" (i.e. that they'll naturally fuel up and compact out on
their own, we've seen this start to happen on staging), but adding a
hammer here in case we need it.

Closes #29357