catalog: Optimize transaction update and delete
This commit adds the following methods to durable catalog transactions:

  - `update_by_key`
  - `update_by_keys`
  - `delete_by_key`
  - `delete_by_keys`

These methods allow callers to update/delete entries by passing in one
or more keys for the modified entries. The existing methods, `update`
and `delete`, use a closure that's passed in as an argument. That
closure is run on every single entry and returns some result that
indicates whether the entry should be updated/deleted. That is
needlessly expensive since we almost always know the keys of the
entries to modify apriori. The new methods can directly modify the
desired entries, without having to check all entries.

This commit also updates the majority of `update` and `delete` callers
to use the new methods. The remaining callers could also be ported over
with some elbow grease, but that is left as future work.

Works towards resolving #28722