compute: make frontier introspection a maintenance task
This commit makes frontier introspection in the compute controller a
"maintenance" task. This means that instead of reporting frontier
changes immediately when they occur, we now only report them once per
second. This is the same cadence with which collected introspection
updates are sent to storage, so the change is not user-visible. It is
desirable for two reasons though:

 * If frontiers change more rapidly than once per second, the old
   approach could emit superfluous introspection updates that would be
   retracted before they reached storage. The new approach avoids that.
 * When we add second-by-second wallclock lag introspection, having the
   frontier introspection update code getting invoked every second will
   come in handy.

Currently, the only frontier introspection emitted by the compute
controller is `ComputeMaterializedViewRefreshes`. This will change in
subsequent commits though.

This commit also cleans up `CollectionIntrospection` a bit, without
making any functional changes.