compute: send frontier introspection data from `Instance`
This commit makes the `Instance` controllers directly report (global and
per-replica) frontier introspection data through the `introspection_tx`.
To this end, the `{Replica}CollectionIntrospection` types are extended
to also handle frontier updates.

An important detail is that `CollectionIntrospection` must not report
global frontiers of MVs because their IDs are shared with storage
collections for which the storage controller already reports frontiers,
so reporting different ones from compute would cause confusion.