controller: avoid double-reporting of compute frontiers
Since the compute controller is now reporting collection frontiers
internally, we need to stop including compute frontiers in what's
written down by the storage controller.

This change removes the need for the global `Controller` to be involved
in the frontier reporting. Ideally reporting of storage frontiers should
be internal to the storage controller, but we punt on that for now to
keep the scope reasonable.