Fix legacy path and re-enable the test
This was an interesting test failure! Proptest was finding a sequence of
changes that reliably hit the "slow path" state sync code, which
involves rebuilding a trace from scratch. My recent change to make the
default of trace structured meant that applying diffs also caused the
follower to enable structured-data roundtripping, even if the upstream
code had not, causing the two states to get out of sync.

This code explicitly sets the flag to match the input. (Which is in
practice on this codepath always false.)

Fun fact: the smallest repro for this appears to be eight separate
operatons, including a mix of empty and non-empty appends and
compactions. Property testing's great.