sql: change IGNORE COLUMNS MySQL option to EXCLUDE COLUMNS
We want to align this option with the forthcoming analogous option for
sinks (#27758). In the context of sinks, "EXCLUDE" reads better than
"IGNORE", because the columns aren't wholly ignored. Rather, they're
allowed to be referenced in the PARTITION BY expression and the HEADERS
option--they're just *excluded* from the final value.

EXCLUDE also nicely aligns with the `SELECT ... EXCLUDE` syntax that was
recently added to DuckDB and Snowflake.

For backwards compatibility, the `IGNORE COLUMNS` option for the MySQL
source continues to be accepted as an alias for `EXCLUDE COLUMNS`.