compute: share collection state from `Instance` tasks
This commit introduces a `SharedCollectionState` type, that enables
sharing of a small subset of the `Instance`'s `CollectionState` with the
`ComputeController` using `Arc<Mutex<_>>`. The data shared consists of
collections' read capabilities and write frontiers.

This change allows making `ComputeController::collection_frontiers`
non-async, by extracting the frontiers from the shared state, rather
than sending commands to the `Instance` tasks and waiting for their
responses.

This change also allows making `acquire_read_hold` non-async, which is
done in the next commit.