compute: remove separate `set_subscribe_target_replica` method
Prior to this change, the controller API for creating a replica-targeted
subscribe was:

 * call `create_dataflow` to create the subscribe dataflow
 * call `set_subscribe_target_replica` to point the dataflow at a
   replica

The second call would fail, for sanity reasons, if the subscribe already
had responses received.

This API is annoying for compute controller decoupling, i.e., moving the
`Instance` controllers into separate tasks. Decoupling will allow the
`Instance` to start the subscribe dataflow as soon as the
`create_dataflow` command is received, and there is no guarantee that
the `set_subscribe_target_replica` command reaches the `Instance` task
before the subscribe received its first response.

To avoid this issues, this commit changes the API to pass the target
replica already in the `create_dataflow` call.

For now, the target is ignored for all dataflow exports that are not
subscribes. In the future, we might want to change that to avoid running
the dataflows of replica-targeted slow-path peeks on replicas that are
not targeted by the peek.