Only include the type info at the top level
Arrow types are recursive, and they're present at all levels of a
hierarchical data type. For example, if I have a struct array, I'll have
a list of fields which include the data types of each child... and a
separate list of children, each of which track their own datatype.

In general this doesn't explode in size for arrow because many types
contain `Arc`s and can be shared. But we lose this sharing when encoding
to proto.

This changes our encoding to pass an "expected" type to encoding and
decoding. Encoding is allowed to omit the type if it matches the
expected type, and decoding is allowed to assume an omitted type is the
expected type. In practice this means that we only need to include the
type info at the top level.

This can be very significant for small arrays. For example, some
single-row arrays I've been testing have their size cut down by about a
third.