Report hydration before frontiers
Report the hydration status before reporting frontiers. Reporting
frontiers also updates the collection frontiers, which can advance the
frontiers to the empty frontier. If reporting hydration status
afterwards, we might observe that the collection has advanced to the
empty frontier and hence suppress the hydration update. The fix is to
first report hydration status before reporting frontiers.

Fixes #29465.

This is an interim fix until the fix for #26682 changes this
implementation again.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>