coord,controller: adjust coordinator state dumping
`ComputeController::dump` is now an `async` method because it requires
waiting for the `Instance` tasks to send back their dumped state. This
commit propagates the `async`-ness up through the coordinator. Dumping
the coordinator state might now block waiting for the compute
controller, but that seems fine given that this mechanism is a rarely
used debugging tool.