coord,controller: adjust hydration checks
The two hydration-check methods of the compute controller,
`clusters_hydrated` and `collections_hydrated_for_replicas`, are now
`async`. This commit adjusts the coordinator to deal with that change.

`clusters_hydrated` is only used for hydration checking in read-only
mode, as part of zero-downtime upgrades. In read-only mode, the
environment doesn't receive user queries, so we don't require low query
latency. Thus, simply awaiting this method is fine.

`collections_hydrated_for_replicas` is invoked as part of `ALTER
CLUSTER` sequencing, so we really don't want to block the coordinator
task awaiting it. Instead we use the `StageResult` mechanism to do the
awaiting off the coordinator task.