compute: acquire read holds in `ComputeController`
This commit moves acquiring of collection read holds from the `Instance`
tasks into the `ComputeController`, using the `SharedCollectionState`
introduced in the previous commit.

In addition to making `ComputeController::acquire_read_hold`
non-`async`, this change also allows us to do the same to
`create_dataflow` and `peek`. For the latter two, we can now perform all
validation in the `ComputeController` and install the required read
holds to ensure that the command will not fail in the `Instance` task
later.