persist: hook up dropped metric for schema cache, too
I only hooked it up for the migrations cache initially because my mental
model was that the schema cache would be tied to the process lifetime.
But, of course, that's only true if there's a long lived handle. Hooking
this up allows us to correct the graph for the total number of entries
in the cache.