adapter: extract an OptimizerCatalog trait
This will help with continual tasks, which need to inject an entry for
the CT itself before it exists in the catalog, but also it gets us one
step closer to pulling the optimizer into its own crate.

I also switched over a few more `adapter::optimize::*::Optimizer`
structs, but had to leave a few because it gets tricky with
`ExprPrepStyle::OneShot` and `eval_unmaterializable_func`.

Touches #28899