persist: Always update opaque value on success
`StateCollections::compare_and_downgrade_since` would return `Ok` if
the shard is finalized or the expected opaque value matches the actual
opaque value. However, the opaque value was only updated to the new
opaque value iff the expected opaque value matched the actual opaque
value, AND the shard was not finalized, AND the actual since was less
than or equal to the new since.

Many callers, including `SinceHandle::compare_and_downgrade_since`,
assumed that if `StateCollections::compare_and_downgrade_since`
returned `Ok`, then the opaque value would be updated. This assumption
is not always true, as described above.

This commit updates `StateCollections::compare_and_downgrade_since`, so
that the opaque value is always updated if the expected opaque value
matches the actual opaque value and the shard is not finalized,
regardless of the since.

The opaque value is still NOT updated if the shard is finalized, which
may still break some callers expectations.

Maybe works towards resolving #29432