0dt: don't check hydration status when using caught-up checks
The newer are-we-caught-up-checks already perform the equivalent of an
are-we-hydrated check for cases where we don't have frontier information
from the currently running deployment. See
https://github.com/MaterializeInc/materialize/blob/250a4db00626f3578bb72059e041bb7bdb964de5/src/compute-client/src/controller/instance.rs#L689.

We initially used are-we-hydrated in addition to are-we-caught-up for
cases where we didn't have frontier information, say when a collection
was migrated/new, but that is now not necessary anymore.