adapter,controller: remove AllowWrites command
As our 0dt deployment thinking has evolved, we no longer plan to support
transitioning a running `environmentd` process from read-only mode to
read-write mode via an externally issued `AllowWrites` command. Instead,
to facilitate the transition, we stop the read-only mode environmentd
process and start a fresh environmentd process in read-write mode.

So, this commit removes the partial implementation of `AllowWrites` at
the adapter/controller level. This will simplify the forthcoming
implementation of 0dt for upsert sources, as we won't need the storage
controller to keep track of which ingestions were started in read-only
mode and which were not.