adapter,storage-controller: lay down 0dt scaffolding for sources
Add a new feature flag called `enable_0dt_deployment_sources` which
controls whether 0dt deployment is attempted for source types that
support it.

To start, only Kafka sources will support 0dt deployments. The rationale
is twofold:

  * Kafka does not have a notion of replication slots, so it's easy to
    have multiple processes reading from the same Kafka topic.

  * Only Kafka sources support the upsert envelope, which is the only
    source operator that has a meaningful rehydration time. Other source
    types (e.g., PostgreSQL) effectively already support 0dt upgrades
    because they can do a cold restart from where they left off nearly
    instantaneously.

This commit does not begin to tackle the hard parts of 0dt deployments
for Kafka upsert sources (read-only mode for the persist sink,
self-correcting upsert state, etc.). Enabling the
`enable_0dt_deployment_sources` is not yet expected to work. The idea is
just to lay down the scaffolding so that we can start to fill in the
missing pieces in parallel.

Touches #27413.