storage: rely on batch truncation in persist_sink
Take a step towards 0dt/m2 and active replication for sources by using
`append` instead of `compare_and_append` in the storage persist sink.

The idea is that, when appending batches in the storage persist sink, if
we discover that the upper is ahead of what we expected, we can simply
discard the updates that are behind the current upper, as we can trust
that the other replica of the source correctly produced those updates.

Note that this needs to be pared with the self-correcting upsert state
(database-issues#8585) to be correct.

Touches database-issues#8095.