storage: retry failed CaA in persist_sink
Take a step towards 0dt/m2 and active replication for sources by
retrying failed `compare_and_append` operations in the storage persist
sink, rather than restarting the whole dataflow.

The idea is that, when appending batches in the storage persist sink, if
we discover that the upper is ahead of what we expected, we can simply
discard the batches whose data is behind current upper, as we can trust
that the other replica of the source correctly produced those updates.

For expediency, this further bakes in the implementation choice that
each batch produced by the storage persist sink contains data at a
single timestamp. In the future, we'll refactor this to rely on
persist's built-in support for batch truncation, but at the moment this
is a code path that's untested in production and is therefore best
avoided.

Note that this needs to be paired with the self-correcting upsert state
(MaterializeInc/database-issues#8585) to be correct.

Alternative to #29724.
Touches MaterializeInc/database-issues#8095.