catalog: Remove deploy gen req from savepoint
Previously, the catalog required a deploy generation when opening in
savepoint mode. However, the upgrade checker has no access to a deploy
generation and was failing to open a savepoint catalog. Ideally, the
upgrade checker would have a deploy generation so that it can test the
exact code path that an upgrade would take. For now, we're removing the
requirement to unblock the upgrade checker.