Retire peeks independently of the coordinator thread
Change how we forward peek responses from the compute instance to the task
handling the client's request. In the past, the peek needed to traverse
the coordinator's main task before it could be sent to the client. Now, it
is sent directly, and only the life cycle logging still happens on the
coordinator main thread.

Related to #29726

Signed-off-by: Moritz Hoffmann <mh@materialize.com>