adapter: plumb a GlobalId into parse_item
ContinualTasks are self-referential. We make this work by using a
placeholder `LocalId` for the CT itself through name resolution and
planning. Then we fill in the real `GlobalId` before constructing this
catalog item. It's nice to be able to do this in one place and
parse_item does the plan -> item work.