Box SelectPlan::select
The SelectPlan struct optionally stores a copy of the select plan. The
SelectStatement struct is rather large (>2KiB), and because we're using
the SelectPlan struct in async state machines, we multiply the space
required to store the statement. This change boxes the select statement
so it doesn't need space on the stack.

A better implementation would be to conditionally store the statement if
it'll be needed based on the explains enabled. This is not as easy as
we're creating the plan before determining what explain modes are enabled.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>