ct: add initial `mz_cluster_replica_metrics_history_ct` builtin
Incrementally roll out CTs by starting with a builtin responsible for
the retention history of `mz_cluster_replica_metrics_history`. This is
modeled as a CT that uses it as an input to maintain a truncated
`mz_cluster_replica_metrics_history_ct` object.

Put the creation of this CT on boot behind a dyncfg, so we can roll it
out slowly and turn it off if anything surprising happens. This means
that we maintain `mz_cluster_replica_metrics_history_ct` but nothing
reads it yet.

We'll want to do an audit of `TODO(ct)`s before turning this on anywhere
important.

Touches materializeinc/database-issues#8427