sinks/kafka: use ThreadedProducer to regularly poll librdkafka
As part of an earlier refactor of mine I switched Kafka sinks to the
`BaseProducer` flavor of client which does not automatically poll the
`librdkafka` client, a crucial task to ensure that messages like
statistics don't pile up. At the same time I didn't ensure poll was
called so a slow memory leak was introduced.

The leak is too slow to notice but we finally did notice and so this PR
fixes this by switching back to the `ThreadedProducer` that spins up a
background thread to take care of regular polls.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>