[compute] refactor LirId treatment (#29837)
In preparation for building the introspection views for attribution, I'm
refactoring how we store `LirId`s to make things a bit more uniform.

### Motivation

   * This PR refactors existing code.

LIR `Plan`s used to just be an enum where every variant had an `lir_id`
field; a `Plan::lir_id()` would pull it out from the appropriate
structure. This PR refactors the code so that a `Plan` is a struct
holding a `PlanNode` (the old enum) and an `LirId`.