storage-controller: flatten DataSourceOther
The `DataSource::Other(DataSourceOther)` variant was meant to house all
data sources that were not "managed" by the storage controller. However:

  * Collections with a data source of `DataSourceOther::TableWrites` are
    *very much* managed by the storage controller. They have a special
    background worker that handles registering them with the txns system
    and incoming write requests.

  * Collections with a data source of `DataSourceOther::Compute` are
    indeed unmanaged by the storage controller, but the name is overly
    restrictive. Collections registered as `DataSourceOther::Compute`
    could just as well be managed by adapter or any other component of
    the system.

So this commit does away with the `DataSourceOther` type. Instead, we
add a top-level `DataSource::Table` variant, for the collections that
need to be managed as tables, and `DataSource::Other`, for the
collections that the storage controller truly should leave alone, and do
not require special handling of any sort.