catalog,persist: add migration for catalog critical since handle
The critical since handle that was previously owned by the catalog will
now be owned by the controller, but the controller uses a different
opaque type for the handle. Add a migration to expire the old critical
since handle created by the catalog so that it can be reregistered with
the correct opaque type by the storage controller.

Without this migration, when upgrading from v0.119, the storage
controller panics during bootstrapping when calling
compare_and_downgrade_since on the catalog shard.