sql,storage: allow using PARTITION BY with ENVELOPE DEBEZIUM
We initially disallowed mixing PARTITION BY with ENVELOPE DEBEZIUM
because, when a message representing an update is processed, if columns
referenced in the PARTITION BY expression change, it's not clear which
partition to send the message to: the PARTITION BY expression could
plausibly be computed on either the `before` values or the `after`
values for the update.

However, it *is* clear which partition to send the message to if the
PARTITION BY expression only refers to columns in the sink's key.  The
key columns are, by definition, the same in both the `before` and the
`after`.

So, this commit allows this behavior. See the docs and tests enclosed in
the patch for details.