pgwire: eagerly flush after writing *all* errors
This is an improvement to #29250. There are a few places where we call
`self.send(BackendMessage::ErrorResponse(..))` without going through
`self.error`. This commit moves the code that flushes after sending an
error so that it's impossible to ever send an error response without
flushing afterwards.

Plausibly related to MaterializeInc/incidents-and-escalations#92.