adapter: Update builtin tables in background
Previously, the adapter would block startup waiting for all updates to
builtin tables to complete. This commit moves this process to the
background and does not block waiting for the writes to complete.

However, we do block on applying the write timestamp with the timestamp
oracle. So anyone in Strict Serializable mode will have to wait until
the writes are complete to read the builtin tables. If someone is in
Serializable mode, they may be able to view the table in a state from
the previous version. This is technically correct, though maybe
slightly undesirable.

Works towards resolving #database-issues/issues/8384