ct: add hack to set an initial since for the builtin CT collection
Hack to make as_of selection happy. It assumes that anything with a zero
upper has the since set to the initial as_of. Ideally we'd write this
down in the durable catalog, but that's hard because of boot ordering
and it's possible that a better long-term fix would be in as_of
selection itself.

Touches MaterializeInc/database-issues#8631