ct: impl CREATE CONTINUAL TASK .. TRANSFORM .. USING as syntactic sugar
The parser learns to recognize `CREATE CONTINUAL TASK .. TRANSFORM ..
USING (..)` and de-sugars it into a full `CREATE CONTINUAL TASK .. AS`
statement. This allows us to incrementally ship a smaller footprint than
the full generality of the feature. The statement remembers how it was
typed in and we can have separate feature flags, but otherwise planning
etc only see this as a normal CT.