catalog: Pass timestamps for transaction commit
This commit adds plumbing for the catalog to accept a lower bound on a
commit timestamp used for transactions. The catalog will ensure that
the transaction is committed at a timestamp greater than or equal to
the provided timestamp.

In general, it is up to the adapter to provide that timestamp.
Currently, the adapter passes in whatever timestamp happens to be lying
around on the `EpochMillis` timeline or the current read timestamp if
one isn't readily available.

The goal of this PR is to move the catalog closer to the `EpochMillis`
timeline and set up the necessary infrastructure to pass timestamps
from the adapter to the catalog.

This work doesn't fully move the catalog onto the `EpochMillis`
timeline, the following work still remains:

  - Always pass a timestamp oracle write timestamp as the commit timestamp.
  - Use the passed commit timestamp exactly in the catalog, instead of
    using it as a lower bound.
  - Always apply the commit timestamp with the timestamp oracle after
    committing.
  - Add the catalog shard to the txn shard with all other tables.
    Otherwise, applying the write timestamp would increase the next
    read timestamp, which would block reads on tables until they were
    advanced.

Works towards resolving #8384