compute: send peek response in `finish_peek`
This commit renames the `remove_peek` method to `finish_peek`, to
reflect that it does more than just removing the peek state. It then
makes sending a peek response part of that method too. This ensures that
a peek never gets removed without being responded to and simplifies the
caller code a bit.