compute,ct: don't report a compute probe, our output_frontier==write_frontier
This was causing problems, in 0dt tests, where we only have a read-only
deployment running. The probe was reporting a frontier based on the
_input_, not actually the write frontier as it was known to the ct_sink.

I believe there is still another curiosity to figure out: why is the
frontier of to_append stuck at `[0]`? Might be that its `persist_source`
is started with an as_of at which no snapshot is available yet. But I
think the fix is correct regardless of that.