Replica expiration applied in controller
Apply the replica expiration only once per replica. Capture the value in
the replica task and encode it in the create-instance command. We only send
this command once per replica. Reconciliation will cause a replica restart
if the value changes and the controller restarts.

This fixes a bug where we could not apply the configuration localized to
a specific replica.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>