storage: fix frontier checking in mint operator
We want the mint operation to be a no-op if either new_from_upper or
binding_ts are not beyond the current source_upper and upper
respectively.

Before, we only checked the former which meant that if the requested
binding_ts was in the past we would attempt to write to the remap shard
updates that are not beyond the upper.

Co-authored-by: Petros Angelatos <petrosagg@gmail.com>
Signed-off-by: Petros Angelatos <petrosagg@gmail.com>