ct: fix bugs in frontier reporting
We we running the `step_backward` operator before
`suppress_early_progress`, which made the frontiers not directly
comparable to the `as_of` given to `suppress_early_progress`. Similarly,
we were reporting incorrect input frontiers (no idea what effect this
has).

This is the easy fix (swap `step_backward` to be later), but adding this
this to the increasing pile of evidence against the weirdness of the
`T-1` idea. While Jan was out, I had an idea for how to make this all
[more normal], but I've been holding off on doing anything it until he
was back. Probably time to hash that out, but since this fix is so
simple, I'd love to get it merged in the interim.

[more normal]: https://materializeinc.slack.com/archives/C075B00521K/p1727808293914569