ct: fix panic when input is not referenced
It's unclear how valuable this is, we could also reasonably return an
error instead of panic. We can still decide to do the planning error too
later, but compute shouldn't panic when presented with a valid dataflow
description. The alternative would be to declare a dataflow description
for a CT that doesn't reference any inputs invalid, but doing that in
the type system seems difficult and doing it in a comment seems brittle.