bin/environmentd: massage connection to metadata database
Use pg8000 rather than psql to connect to the metadata database. This
gives us more control over the SQL connection, and allows us to easily
detect whether we're connected to a real PostgreSQL server or
CockroachDB.

Also, don't attempt to create a database when running against real
PostgreSQL. PostgreSQL requires you to specify an extant database to
establish the connection in the first place, so there's only pain to be
had by trying to create the requested database, as it requires guessing
the name of a database that does exist *and* a user that has access to
that database, and every system has a different convention here as to
what PostgreSQL databases are guaranteed to exist (e.g., a database
named after the current user or a database for the superuser
"postgres").