flexible-deployments: include orchestratord updates (#29964)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation
Updating the Helm chart boilerplate to include the initial
implementation of orchestrated as per:
https://github.com/MaterializeInc/materialize/pull/29922

This is still mostly a boilerplate. Still needs work and testing. And as
discussed we will be changing CRDB with Postgres.